resolved issue with recoding, filtered out "Other"
Filtering out "Other" wasn't working because there were errors in our recoding. I fixed these and now it works! So the clean_data object has all observations with "Other" removed. (Not the right choice in terms of actual stats work, but good practice with data cleaning, and simplifies our results for this assignment)